Added subdomain column to SQLite DB, also added\n some rules for shoudVisit() to avoid informatics and calendar

+++ src/CrawlStat.java
@@ -29,7 +29,7 @@ import edu.uci.ics.crawler4j.url.WebURL;
 
 public class CrawlStat {
         private static int 				mNumPages;			//# of pages
-        private static long 			mLongestPageLength; //# words in longest page
+        private static int 			mLongestPageLength; //# words in longest page
         private static Map<String, Integer> 	mSubdomainsMap;		//subdomain names and frequencies
         private static Map<String, Integer> 	mWordMap;			//words and frequencies Map
         private static Map<String, Integer> 	m2gramMap;			//2grams and frequencies Map
@@ -98,11 +98,11 @@ public class CrawlStat {
         	mNumPages += 1;
         }
         
-        public long getLongestPageLength() {
+        public int getLongestPageLength() {
             return mLongestPageLength;
         }
         
-        public void setLongestPageLength(long pageLength) {
+        public void setLongestPageLength(int pageLength) {
         	if (pageLength > mLongestPageLength) {
         		mLongestPageLength = pageLength;
         	}
